pyproject.toml
[project]
name = "radiosondy-landed-waypoints"
version = "0.1.0"
description = "A Python script to extract radiosonde landing coordinates and generate a GPX file."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "gpxpy>=1.6.2",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "requests>=2.32.5",
    "telethon>=1.41.2",
]
